fix: project selector in header takes time to load (#21)
